% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InternalFunctions.R
\name{.Diff}
\alias{.Diff}
\title{Internal functions}
\usage{
.Diff(x, lag = 1, ...)
}
\arguments{
\item{x}{time series}

\item{lag}{which lag}

\item{...}{further parameters given to diff()
Functions used internally in the seasonality package}
}
\description{
Get differenced series
}
\author{
Daniel Ollech
}
