% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InternalFunctions.R
\name{.Lag}
\alias{.Lag}
\title{Internal functions}
\usage{
.Lag(x, k)
}
\arguments{
\item{x}{time series}

\item{k}{number of lags
Functions used internally in the seasonality package}
}
\description{
Get lag
}
\author{
Daniel Ollech
}
