% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fried.R
\name{fried}
\alias{fried}
\title{Friedman Rank test}
\usage{
fried(x, freq = NA, diff = T, residuals = F, autoarima = T)
}
\arguments{
\item{x}{time series}

\item{freq}{Frequency of the time series}

\item{diff}{Shall the differenced series be tested?}

\item{residuals}{Shall the residuals of an ARIMA model be tested?}

\item{autoarima}{Use automatic instead of a (0,1,1) ARIMA model?}
}
\description{
Test for seasonality in a time series.
}
\details{
If residuals=FALSE the autoarima settings are ignored.

If residuals=TRUE, a non-seasonal ARIMA model is estimated for the time series. And the residuals of the fitted model are used as input to the test statistic. If an automatic order selection is used, the Hyndman-Khandakar algorithm is employed with max(p)=max(q) <= 3.
}
\examples{
fried(ts(rnorm(120, 10,10), frequency=12))
fried(ts(rnorm(1200, 10,10), frequency=7))
}
\references{
Friedman, M. (1937). The Use of Ranks to Avoid the Assumption of Normality Implicit in the Analysis of Variance. Journal of the American Statistical Association 32 (200), 675-701.

Hyndman, R. J. and Y. Khandakar (2008). Automatic Time Series Forecasting: The forecast Package for R. Journal of Statistical Software 27 (3), 1-22.

Ollech, D. and Webel, K. (2020). A random forest-based approach to identifying the most informative seasonality tests. Deutsche Bundesbank's Discussion Paper series 55/2020.
}
\author{
Daniel Ollech
}
