% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_seastests.R
\name{print.seastests}
\alias{print.seastests}
\title{Generic function for class seastests}
\usage{
\method{print}{seastests}(x, ...)
}
\arguments{
\item{x}{result from seasonality test}

\item{...}{additional arguments}
}
\description{
Generic function for class seastests
}
\examples{
a <- qs(ts(rnorm(120, 10,10), frequency=12))
print(a)
summary(a)
}
\author{
Daniel Ollech
}
