% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_seasinttests.R
\name{summary.seasinttests}
\alias{summary.seasinttests}
\title{Generic functions for class seasinttests}
\usage{
\method{summary}{seasinttests}(object, ...)
}
\arguments{
\item{object}{result from seasonal integration test}

\item{...}{additional arguments}
}
\description{
Generic functions for class seasinttests
}
\examples{
a <- qs(ts(rnorm(120, 10,10), frequency=12))
print(a)
summary(a)
}
\author{
Daniel Ollech
}
